cleanup: remove junk files, organize SQL, update gitignore
Removed from tracking:
- NROTFT_old.jar, done.jar (build artifacts)
- backup_*.sql (database backups)
- tmp_bg_preview/ (temp preview images)
- CheckPlayer.java, DebugEnc.java, FixEncoding.java, TestDB.java (debug utilities)
- find_ips.js, patch.js, patch.ps1 (temp scripts)
- compile_log.txt, sources.txt, blocked_ips.txt (runtime files)
- META-INF/, nbproject/build-impl.xml~ (build artifacts)

Organized:
- Moved 8 SQL scripts from root -> sql/ folder

Fixed:
- ChuyenKhoanManager.java: cast double to long for notifyRecharge()
- Updated .gitignore with comprehensive rules

+++ src/nro/server/ChuyenKhoanManager.java
@@ -525,7 +525,7 @@ public class ChuyenKhoanManager {
                     try {
                         NotificationService.gI().notifyRecharge(
                             player.name != null ? player.name : "ID:" + player.id,
-                            transaction.amount, "ATM Bank (Auto)");
+                            (long) transaction.amount, "ATM Bank (Auto)");
                     } catch (Exception ignored) {}
                     Service.gI().sendThongBao(player, "Bạn nhận được tiền là: " + (transaction.amount * 10) + " và thưởng: " + bonus);
